pyproject.toml
[project]
name = "enrollments-api"
version = "1"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]==0.115.12",
    "pymongo==4.12.1",
    "mongomock==4.3.0",
    "pydantic-settings==2.9.1",
    "requests==2.32.3",
    "pika==1.3.2",
    "pip-tools"
]

[tool.setuptools]
py-modules = []
